Remove duplicated code from sandbox
