+++ lib/elixir_console/sandbox/erlang_modules_absence.ex
@@ -3,10 +3,8 @@ defmodule ElixirConsole.Sandbox.ErlangModulesAbsence do
   Analyze the AST and check if Erlang modules are not present.
   """
 
-  alias ElixirConsole.Sandbox.CommandValidator
+  alias ElixirConsole.Sandbox.{CommandValidator, Util}
   @behaviour CommandValidator
-
-  @az_range 97..122
 
   @impl CommandValidator
   def validate(ast) do
@@ -28,7 +26,7 @@ defmodule ElixirConsole.Sandbox.ErlangModulesAbsence do
   end
 
   defp valid?({:., _, [module, _]} = elem, acc) do
-    if is_erlang_module?(module) do
+    if Util.is_erlang_module?(module) do
       {elem, [{:error, module} | acc]}
     else
       {elem, acc}
@@ -36,15 +34,4 @@ defmodule ElixirConsole.Sandbox.ErlangModulesAbsence do
   end
 
   defp valid?(elem, acc), do: {elem, acc}
-
-  defp is_erlang_module?(module) when not is_atom(module), do: false
-
-  defp is_erlang_module?(module) do
-    module
-    |> to_charlist
-    |> starts_with_lowercase?
-  end
-
-  defp starts_with_lowercase?([first_char | _]) when first_char in @az_range, do: true
-  defp starts_with_lowercase?(_module_charlist), do: false
 end
